Rename `write.py` into `zarr.py`
